`size_t` feels more descriptive for the purpose of the variable.

+++ exd.c
@@ -32,7 +32,7 @@
 
 int main (int argc, char *argv[])
 {
-    long address;
+    size_t address;
     FILE *fd;
 
     int rowWasZero;
